Revert "tentative changes to the function and class definition, need testing in lab to see the most efficient usages for the function/constructor"
This reverts commit a97c0848d7e9be1468ecf2f923597f82e298de49.

+++ classes/Rubisco_Chain.m
@@ -10,7 +10,6 @@ classdef Rubisco_Chain < handle
         tag = nan;
         tags uint32 = uint32([]); % tags of rubiscos in chain
         indices uint32 = uint32([]); % indices of rubiscos in chain
-        % Primarily testing two ways - tags using casting initialization, indices using direct initialization in the constructor
         average_vector = []; % average vector of rubiscos in chain
         centroid = []; % centroid of chain
         eigenvalues = []; % eigenvalues of the S tensor
@@ -28,7 +27,7 @@ classdef Rubisco_Chain < handle
                 chain.reg = reg;
                 chain.tags = tags;
                 chain.length = length(tags);
-                chain.indices = uint32(indices);
+                chain.indices = indices;
                 chain.type = type;
                 chain.index = index;
                 chain.tag = tag;
